+++ datagen/src/main/java/io/github/jamalam360/utility_belt/datagen/Tags.java
@@ -7,6 +7,7 @@ import net.minecraft.core.HolderLookup;
 import net.minecraft.core.registries.Registries;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.tags.TagKey;
+import net.minecraft.world.level.block.Blocks;
 
 import java.util.concurrent.CompletableFuture;